requirements.txt
python-dotenv
langchain
langchain_community
langchain_openai
langchain_chroma
Flask
Flask-Cors
flask-talisman
Markdown
html2text
python-markdown-math
python-pptx